+++ airbyte-integrations/connectors/source-postgres/src/main/kotlin/io/airbyte/integrations/source/postgres/cdc/PostgresSourceCdcMetaFields.kt
@@ -1,0 +1,19 @@
+/*
+ * Copyright (c) 2025 Airbyte, Inc., all rights reserved.
+ */
+
+package io.airbyte.integrations.source.postgres.cdc
+
+import io.airbyte.cdk.discover.CdcStringMetaFieldType
+import io.airbyte.cdk.discover.FieldType
+import io.airbyte.cdk.discover.MetaField
+
+enum class PostgresSourceCdcMetaFields(
+    override val type: FieldType,
+) : MetaField {
+    CDC_LSN(CdcStringMetaFieldType),
+    ;
+
+    override val id: String
+        get() = MetaField.META_PREFIX + name.lowercase()
+}