+++ airbyte-cdk/bulk/changelog.md
@@ -1,3 +1,7 @@
+## WIP changes for Postgres (version number TBD)
+
+* Field replaced by EmittedField
+* CDC Position objects now implement PartiallyOrdered rather than Comparable
 ## Version 0.1.89
 
 load cdk: components tests: data coercion tests for int+number